MS-13 플레이어 이동 동기화

+++ Server/GameServer/game/ms_coordinate.h
@@ -2,8 +2,8 @@
 #include <cstdint>
 
 struct MsCoordinate {
-  int32_t x;
-  int32_t y;
+  double x;
+  double y;
   int16_t grid_x;
   int16_t grid_y;
   int16_t grid_size = 1;
